Move gameplay modules into structured folders
