+++ frontend/src/systems/inventory-system.js
